source/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
pyjwt